consumer-prediction/requirements.txt
kafka-python
pymongo
python-dotenv
requests

